Priliminary de-macro prep and cleanup
