+++ Cargo.toml
@@ -14,6 +14,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 gcd = ">=2.1,<3.0"
+num-traits = { version = "0.2.19", default-features = false }
 typenum = "1.17.0"
 
 [dependencies.defmt]
